feat: Провести аудит на соответствие ТЗ и внести первоочередные исправления
Создан документ `AUDIT_FINDINGS.md`, детально описывающий все несоответствия между кодовой базой и техническим заданием.

На основе аудита внесены следующие первоочередные исправления:

1.  **Исправлены Модели и Схемы:**
    *   Тип поля `shifts_count` в модели `Payroll` изменен на `Integer`.
    *   В схему ответа `CouponRedeemResponse` добавлен объект клиента для соответствия спецификации API.

2.  **Реализована Бизнес-Логика в `RedemptionService`:**
    *   Внедрена система структурированных исключений для API.
    *   Добавлены проверки `min_purchase` и условий (`conditions`) из шаблона купона.
    *   Реализована поддержка многоразовых купонов (проверки `usage_limit` и `per_user_limit`).
    *   Добавлен каркас для логики стэкинга скидок.
    *   Реализован сценарий "покупка без купона".
    *   Исправлена логика проверки для общих (неперсональных) купонов.

3.  **Доработан `CouponService`:**
    *   Добавлена проверка на соответствие клиента условиям шаблона при выдаче купона.
    *   Реализован расчет срока действия купона на основе `expiration_days` из шаблона.

+++ src/app/db/models/hr.py
@@ -75,7 +75,7 @@ class Payroll(Base):
     )
     month: Mapped[Date] = mapped_column(Date, nullable=False)
     shifts_count: Mapped[int] = mapped_column(
-        Numeric(12, 2), nullable=False, server_default="0"
+        Integer, nullable=False, server_default="0"
     )
     gross: Mapped[float] = mapped_column(
         Numeric(12, 2), nullable=False, server_default="0"
